+++ contracts/java/src/org/iota/wasp/wasmlib/bytes/BytesEncoder.java
@@ -9,67 +9,67 @@ import java.io.*;
 import java.nio.charset.*;
 
 public class BytesEncoder {
-	ByteArrayOutputStream data;
+    ByteArrayOutputStream data;
 
-	public BytesEncoder() {
-		data = new ByteArrayOutputStream();
-	}
+    public BytesEncoder() {
+        data = new ByteArrayOutputStream();
+    }
 
-	public BytesEncoder Address(ScAddress value) {
-		return Bytes(value.toBytes());
-	}
+    public BytesEncoder Address(ScAddress value) {
+        return Bytes(value.toBytes());
+    }
 
-	public BytesEncoder AgentId(ScAgentId value) {
-		return Bytes(value.toBytes());
-	}
+    public BytesEncoder AgentId(ScAgentId value) {
+        return Bytes(value.toBytes());
+    }
 
-	public BytesEncoder Bytes(byte[] value) {
-		Int(value.length);
-		try {
-			data.write(value);
-		} catch (IOException e) {
-		}
-		return this;
-	}
+    public BytesEncoder Bytes(byte[] value) {
+        Int64(value.length);
+        try {
+            data.write(value);
+        } catch (IOException e) {
+        }
+        return this;
+    }
 
-	public BytesEncoder ChainId(ScChainId value) {
-		return Bytes(value.toBytes());
-	}
+    public BytesEncoder ChainId(ScChainId value) {
+        return Bytes(value.toBytes());
+    }
 
-	public BytesEncoder Color(ScColor value) {
-		return Bytes(value.toBytes());
-	}
+    public BytesEncoder Color(ScColor value) {
+        return Bytes(value.toBytes());
+    }
 
-	public BytesEncoder ContractId(ScContractId value) {
-		return Bytes(value.toBytes());
-	}
+    public BytesEncoder ContractId(ScContractId value) {
+        return Bytes(value.toBytes());
+    }
 
-	public byte[] Data() {
-		return data.toByteArray();
-	}
+    public byte[] Data() {
+        return data.toByteArray();
+    }
 
-	public BytesEncoder Hash(ScHash value) {
-		return Bytes(value.toBytes());
-	}
+    public BytesEncoder Hash(ScHash value) {
+        return Bytes(value.toBytes());
+    }
 
-	public BytesEncoder Hname(ScHname value) {
-		return Bytes(value.toBytes());
-	}
+    public BytesEncoder Hname(ScHname value) {
+        return Bytes(value.toBytes());
+    }
 
-	public BytesEncoder Int(long value) {
-		for (; ; ) {
-			byte b = (byte) value;
-			byte s = (byte) (b & 0x40);
-			value >>= 7;
-			if ((value == 0 && s == 0) || (value == -1 && s != 0)) {
-				data.write((byte) (b & 0x7f));
-				return this;
-			}
-			data.write((byte) (b | 0x80));
-		}
-	}
+    public BytesEncoder Int64(long value) {
+        for (; ; ) {
+            byte b = (byte) value;
+            byte s = (byte) (b & 0x40);
+            value >>= 7;
+            if ((value == 0 && s == 0) || (value == -1 && s != 0)) {
+                data.write((byte) (b & 0x7f));
+                return this;
+            }
+            data.write((byte) (b | 0x80));
+        }
+    }
 
-	public BytesEncoder String(String value) {
-		return Bytes(value.getBytes(StandardCharsets.UTF_8));
-	}
+    public BytesEncoder String(String value) {
+        return Bytes(value.getBytes(StandardCharsets.UTF_8));
+    }
 }
